Properly collect zombie processes after termination
As it turns out, simply killing the Chromedriver process instance is not
sufficient to release associated resources. Rather, we need to make sure
to actively wait for the process to terminate and read it's exit code in
order to free its remnants. Make sure to do that in order to prevent
resource leakage in long-lived processes repeatedly spawning
Chromedriver instances.